requirements.txt
flask==3.0.3
openai==1.35.7
python-dotenv==1.0.1
gunicorn==22.0.0
